+++ racingcar/model/car.go
@@ -1,4 +1,4 @@
-package racingcar
+package model
 
 import "errors"
