+++ racingcar/model/racingcar.go
@@ -1,4 +1,4 @@
-package racingcar
+package model
 
 // 예시 함수, 구현 시 제거
 func Hello() string {
