+++ racingcar/model/run_result.go
@@ -1,4 +1,4 @@
-package racingcar
+package model
 
 type RunResult struct {
 	CarNameToPosition map[string]int
