+++ test/model/application_test.go
@@ -1,7 +1,7 @@
-package test
+package model
 
 import (
-	"github.com/poi1649/go-racing-car/racingcar"
+	"github.com/poi1649/go-racing-car/racingcar/model"
 	"github.com/stretchr/testify/assert"
 
 	"testing"
@@ -10,7 +10,7 @@ import (
 // 예시 테스트 함수, 구현 시 제거
 func TestHello(t *testing.T) {
 	expected := "Hello, Racing Car!"
-	actual := racingcar.Hello()
+	actual := model.Hello()
 	assert.Equal(
 		t,
 		expected,
