+++ test/model/cars_test.go
@@ -1,35 +1,35 @@
-package test
+package model
 
 import (
-	"github.com/poi1649/go-racing-car/racingcar"
+	"github.com/poi1649/go-racing-car/racingcar/model"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
 func TestGenerateCars(t *testing.T) {
-	car1, _ := racingcar.NewCar("test1")
-	car2, _ := racingcar.NewCar("test2")
-	car3, _ := racingcar.NewCar("test3")
-	cars := racingcar.NewCars([]*racingcar.Car{car1, car2, car3})
+	car1, _ := model.NewCar("test1")
+	car2, _ := model.NewCar("test2")
+	car3, _ := model.NewCar("test3")
+	cars := model.NewCars([]*model.Car{car1, car2, car3})
 	assert.Equal(t, 3, len(cars.Cars))
 }
 
 func TestGenerateCarsWithMethod(t *testing.T) {
-	cars, _ := racingcar.GenerateCars([]string{"test1", "test2", "test3"})
+	cars, _ := model.GenerateCars([]string{"test1", "test2", "test3"})
 	assert.Equal(t, 3, len(cars.Cars))
 }
 
 func TestGenerateCarsWithInvalidName(t *testing.T) {
-	_, err := racingcar.GenerateCars([]string{"test1", "test2", ""})
+	_, err := model.GenerateCars([]string{"test1", "test2", ""})
 	errMsg := "차 이름은 1자 이상 5자 이하만 가능합니다"
 	assert.Equal(t, errMsg, err.Error())
 }
 
 func TestGetWinners(t *testing.T) {
-	car1, _ := racingcar.NewCar("test1")
-	car2, _ := racingcar.NewCar("test2")
-	car3, _ := racingcar.NewCar("test3")
-	cars := racingcar.NewCars([]*racingcar.Car{car1, car2, car3})
+	car1, _ := model.NewCar("test1")
+	car2, _ := model.NewCar("test2")
+	car3, _ := model.NewCar("test3")
+	cars := model.NewCars([]*model.Car{car1, car2, car3})
 	car1.Move(3)
 	car2.Move(5)
 	car3.Move(1)
@@ -39,10 +39,10 @@ func TestGetWinners(t *testing.T) {
 }
 
 func TestGetWinnersWithSamePosition(t *testing.T) {
-	car1, _ := racingcar.NewCar("test1")
-	car2, _ := racingcar.NewCar("test2")
-	car3, _ := racingcar.NewCar("test3")
-	cars := racingcar.NewCars([]*racingcar.Car{car1, car2, car3})
+	car1, _ := model.NewCar("test1")
+	car2, _ := model.NewCar("test2")
+	car3, _ := model.NewCar("test3")
+	cars := model.NewCars([]*model.Car{car1, car2, car3})
 	car1.Move(3)
 	car2.Move(5)
 	car3.Move(5)
